Update for AWS-CLI v2

+++ 3rd-Bootstrap_AmazonLinux-1-HVM.sh
@@ -149,17 +149,22 @@ fi
 #-------------------------------------------------------------------------------
 
 # Package Uninstall AWS-CLI v1 Tools (from RPM Package)
-aws --version
-which aws
+if [ $(compgen -ac | sort | uniq | grep -x aws) ]; then
+	aws --version
 
-yum remove -y awscli
+	which aws
+
+	rpm -qi awscli
+
+	yum remove -y awscli
+fi
 
 # Package download AWS-CLI v2 Tools (from Bundle Installer)
 curl -sS "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "/tmp/awscliv2.zip"
 unzip "/tmp/awscliv2.zip" -d /tmp/
 
 # Package Install AWS-CLI v2 Tools (from Bundle Installer)
-/tmp/aws/install -i "/opt/aws/awscli" -b "/bin"
+/tmp/aws/install -i "/opt/aws/awscli" -b "/usr/bin"
 
 aws --version
 
@@ -184,6 +189,9 @@ __EOF__
 
 # Setting AWS-CLI Logging
 aws configure set cli_history enabled
+
+# Setting AWS-CLI Pager settings
+aws configure set cli_pager ''
 
 # Getting AWS-CLI default Region & Output format
 aws configure list
